removed tar and gzip, added system_dir/tmp, fixed bugs in us_write_config.cpp
Note: There is no explicit check to make sure that dirs defined in etc/usrc.conf
are actually present. We need to rely on this for various files written by 
ultrascan. The check is not performed until us_config_gui is executed! It
should be done each time us is started.