Allow less delay between acceleration end and first scan
The previous calculation relied on an overestimated acceleration rate in the speed step resulting in an offset of -3 for the scan timing based on experimental observations. The updated calculation for the timestate makes this correction obsolete. In order to keep not suddenly display errors for older data, the offset of +1 is applied due to making tf_aend a double instead of an integer avoiding potential rounding issues.

Signed-off-by: doluk <69309597+doluk@users.noreply.github.com>