Remove redundant `setVisible(true)` call in `us_plot3d.cpp`
Virtual functions shouldn't be called in the initializer of a class. Further this will make the window always visible, when we might not want it (us_fematch in auto mode for example directly sets the visible back to false). By removing this we prevent any sort of flashing the window